Refactoring pkg, implement plugin-system
