+++ .gitignore
@@ -6,3 +6,4 @@ logs
 coverage*.out
 integration_test/certs/tls.crt
 integration_test/certs/tls.key
+.env
