+++ Makefile
@@ -4,6 +4,17 @@ ifeq (${DOCKER_COMPOSE},)
 DOCKER_COMPOSE = docker compose
 endif
 
+VERSION    := $(shell git describe --tags --always --dirty 2>/dev/null || echo "dev")
+COMMIT     := $(shell git rev-parse --short HEAD 2>/dev/null || echo "unknown")
+BUILD_DATE := $(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
+LDFLAGS    := -X github.com/mariocandela/beelzebub/v3/cli.Version=$(VERSION) \
+              -X github.com/mariocandela/beelzebub/v3/cli.CommitSHA=$(COMMIT) \
+              -X github.com/mariocandela/beelzebub/v3/cli.BuildDate=$(BUILD_DATE)
+
+.PHONY: build
+build:
+	go build -ldflags "$(LDFLAGS)" -o beelzebub .
+
 .PHONY: beelzebub.start
 beelzebub.start:
 	${DOCKER_COMPOSE} build;
