+++ internal/plugins/maze_test.go
